Consolidate example projects
